slight restructure of the root, v1 and sync commands - got get -u - added support for arm64 builds - prepared badges in readme - added LICENSE
