+++ Makefile
@@ -4,12 +4,18 @@
 BIN = bin
 GOBUILD = go build
 
+simple-build:
+	$(GOBUILD) -o $(BIN)/gitsync
+
 # build the project
 build:
 	$(GOBUILD) -o $(BIN)/gitsync
 	GOOS=windows GOARCH=amd64 $(GOBUILD) -o $(BIN)/gitsync-windows-amd64.exe
+	GOOS=windows GOARCH=arm64 $(GOBUILD) -o $(BIN)/gitsync-windows-arm64.exe
 	GOOS=linux GOARCH=amd64 $(GOBUILD) -o $(BIN)/gitsync-linux-amd64
+	GOOS=linux GOARCH=arm64 $(GOBUILD) -o $(BIN)/gitsync-linux-arm64
 	GOOS=darwin GOARCH=amd64 $(GOBUILD) -o $(BIN)/gitsync-darwin-amd64
+	GOOS=darwin GOARCH=arm64 $(GOBUILD) -o $(BIN)/gitsync-darwin-arm64
 
 # run quality assessment checks
 check:
