+++ README.md
@@ -1,2 +1,9 @@
-# gitsync-cli
+# GitSync CLI
+
+[![github.com release badge](https://img.shields.io/github/release/restechnica/gitsync-cli.svg)](https://github.com/restechnica/gitsync-cli/)
+[![github.com workflow badge](https://github.com/restechnica/gitsync-cli/workflows/main/badge.svg)](https://github.com/restechnica/gitsync-cli/actions?query=workflow%3Amain)
+[![go.pkg.dev badge](https://pkg.go.dev/badge/github.com/restechnica/gitsync-cli)](https://pkg.go.dev/github.com/restechnica/gitsync-cli)
+[![goreportcard.com badge](https://goreportcard.com/badge/github.com/restechnica/gitsync-cli)](https://goreportcard.com/report/github.com/restechnica/gitsync-cli)
+[![img.shields.io MPL2 license badge](https://img.shields.io/github/license/restechnica/gitsync-cli)](./LICENSE)
+
 A CLI to sync git repositories.
