+++ go.mod
@@ -2,7 +2,7 @@ module github.com/restechnica/gitsync-cli
 
 go 1.17
 
-require github.com/spf13/cobra v1.3.0
+require github.com/spf13/cobra v1.4.0
 
 require (
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
