+++ pkg/cli/version.go
@@ -1,4 +1,4 @@
-package v1
+package cli
 
 import (
 	"fmt"
